Move ProblemNotice et. al into its own file
