+++ src/python_testing/TC_DefaultWarnings.py
@@ -38,8 +38,8 @@ from typing import Optional
 
 import chip.clusters as Clusters
 from chip.testing.basic_composition import BasicCompositionTests
-from chip.testing.matter_testing import (AttributePathLocation, ClusterPathLocation, MatterBaseTest, ProblemLocation, ProblemNotice,
-                                         ProblemSeverity, TestStep, async_test_body, default_matter_test_main)
+from chip.testing.matter_testing import MatterBaseTest, TestStep, async_test_body, default_matter_test_main
+from chip.testing.problem_notices import AttributePathLocation, ClusterPathLocation, ProblemLocation, ProblemNotice, ProblemSeverity
 from mobly import asserts
 
 FLAG_PRODUCT_NAME = "pixit_allow_test_in_product_name"
