+++ src/python_testing/matter_testing_infrastructure/BUILD.gn
@@ -48,6 +48,7 @@ pw_python_package("chip-testing-module") {
     "chip/testing/matter_testing.py",
     "chip/testing/metadata.py",
     "chip/testing/pics.py",
+    "chip/testing/problem_notices.py",
     "chip/testing/runner.py",
     "chip/testing/spec_parsing.py",
     "chip/testing/taglist_and_topology_test.py",
