+++ src/python_testing/matter_testing_infrastructure/chip/testing/choice_conformance.py
@@ -1,6 +1,6 @@
 from chip.testing.conformance import Choice, ConformanceDecisionWithChoice
 from chip.testing.global_attribute_ids import GlobalAttributeIds
-from chip.testing.matter_testing import AttributePathLocation, ProblemNotice, ProblemSeverity
+from chip.testing.problem_notices import AttributePathLocation, ProblemNotice, ProblemSeverity
 from chip.testing.spec_parsing import XmlCluster
 from chip.tlv import uint
